added backButton in testresults, moved SubmitButton to components
