+++ src/components/TestIntro.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styles from './TestIntro.module.css';
 import Button from './Button'
